+++ lib/change_health/response/trading_partners_data.rb
@@ -2,7 +2,7 @@ module ChangeHealth
   module Response
     class TradingPartnersData < Array
       def initialize(trading_partners)
-        super(trading_partners.map {|trading_partner| ChangeHealth::Models::TradingPartner.new(trading_partner) })
+        super(trading_partners.map {|trading_partner| ChangeHealth::Response::TradingPartnerData.new(trading_partner) })
       end
     end
   end