+++ test/change_health/models/trading_partner_test.rb
@@ -3,10 +3,10 @@ require 'test_helper'
 class TradingPartnerTest < Minitest::Test
 
   describe "object" do
-    let(:trading_partner) { ChangeHealth::Models::TradingPartner.new(id: '1', name: 'Test Partner') }
+    let(:trading_partner) { ChangeHealth::Models::TradingPartner.new(service_id: '1', name: 'Test Partner') }
 
     it "has id" do
-      assert trading_partner.respond_to?(:id)
+      assert trading_partner.respond_to?(:service_id)
     end
 
     it "has name" do